[v1.1.0] Bug fixes & improvements
Functions that accept URLs and files as parameters (i.e.
api.send_audio()) will now assume that, if the given string is a file
path, the intended action is to upload that file and send it.

Using api.send_audio(message.chat.id, ‘/tmp/song.mp3’) would send the
file song.mp3 in the /tmp/ directory. If that file didn’t exist, then
it would return false.

No conflicts would arise since file paths are structurally different to
the HTTP(S) URLs that the Telegram bot API accepts, so I felt this was
the best move.

Please bare in mind the Telegram bot API still has strict limits as to
what type of files you can upload, and a lot of methods have a maximum
file size restriction of 50 MB (although it’s noted in the official
documentation that this could be changed in the future).